Codegen: emit maps in libbpf style
This is another step towards delegating loading and attachment of BPF
programs to libbpf. libbpf requires maps to be emitted in certain format
with attached BTF information.

The expected format is that each map is represented by a global variable
in the ".maps" section. The variable should be of a struct type with the
following fields:
- "type"        map type (e.g. BPF_MAP_TYPE_HASH)
- "max_entries" maximum number of entries
- "key"         key type
- "value"       value type

The variable and its type should have appropriate BTF entries. This is
done by emitting DWARF debugging information for the map which LLVM
converts to BTF when emitting the program.

It is also necessary to fixup BTF emitted by LLVM b/c it's missing some
information. This is usually done by libbpf's linker but since we're
passing BTF to the BPF program directly (instead of using bpf_object),
we must perform this step manually.

See in-code comments for more implementation details.

For now, only user-defined maps are handled this way. Support for
bpftrace internal maps will follow.