Codegen: emit internal maps in libbpf style
This follows up on the previous commit by emitting bpftrace internal
maps in libbpf style (global variable + BTF). This is done in the same
way for most of the cases, with the only exceptions being the
BPF_MAP_TYPE_RINGBUF map type which doesn't have the 'key' and 'value'
fields.

There is also a change in codegen tests for 'kstack' and 'ustack'. One
of the previous commits updated the names of stack maps to include the
stack mode and limit to be able to differentiate distinct maps by name.
Since the name is now a part of codegen (it's the name of the emit
variable), it's no longer true that 'kstack()' and 'kstack(perf)'
generate the same codegen.